do not use stress test diff. returns top N for CPU/Mem
