+++ .gitlab/source_test/ebpf.yml
@@ -66,15 +66,6 @@ tests_ebpf_x64:
     - inv -e security-agent.build-functional-tests --output=$DD_AGENT_TESTING_DIR/site-cookbooks/dd-security-agent-check/files/testsuite --build-flags "-race"
     # Compile runtime security stress tests to be executed in kitchen tests
     - inv -e security-agent.build-stress-tests --output=$DD_AGENT_TESTING_DIR/site-cookbooks/dd-security-agent-check/files/stresssuite
-    # Compile main version for comparison, uncomment following lines when merged
-    # - git checkout -f main
-    # - git pull
-    # - inv -e deps
-    # - inv -e system-probe.build --bundle-ebpf --incremental-build
-    # - inv -e security-agent.build-stress-tests --output=$DD_AGENT_TESTING_DIR/site-cookbooks/dd-security-agent-check/files/stresssuite-master
-    # - git reset --hard
-    # - git checkout -
-
 
 tests_ebpf_arm64:
   extends: .tests_linux_ebpf
