+++ pkg/security/tests/stresser_test.go
@@ -107,7 +107,8 @@ type StressReport struct {
 		Value float64
 		Unit  string
 	} `json:",omitempty"`
-	Top []byte `json:"-"`
+	TopCPU []byte `json:"-"`
+	TopMem []byte `json:"-"`
 }
 
 // AddMetric add custom metrics to the report
@@ -137,8 +138,8 @@ func (s *StressReport) Delta() float64 {
 }
 
 // Print prints the report in a human readable format
-func (s *StressReport) Print() {
-	fmt.Println("----- Stress Report -----")
+func (s *StressReport) Print(t *testing.T) {
+	fmt.Printf("----- Stress Report for %s -----\n", t.Name())
 	w := tabwriter.NewWriter(os.Stdout, 0, 8, 1, '\t', tabwriter.AlignRight)
 	fmt.Fprintf(w, "%s\t\t%d iterations\t%15.4f ns/iteration", s.Duration, s.Iteration, float64(s.Duration.Nanoseconds())/float64(s.Iteration))
 	if s.Extras != nil {
@@ -155,8 +156,13 @@ func (s *StressReport) Print() {
 	w.Flush()
 
 	fmt.Println()
-	fmt.Println("----- Profiling Report -----")
-	fmt.Println(string(s.Top))
+	fmt.Printf("----- Profiling Report CPU for %s -----\n", t.Name())
+	fmt.Println(string(s.TopCPU))
+	fmt.Println()
+
+	fmt.Println()
+	fmt.Printf("----- Profiling Report Memory for %s -----\n", t.Name())
+	fmt.Println(string(s.TopMem))
 	fmt.Println()
 }
 
@@ -318,7 +324,13 @@ LOOP:
 		}
 	}
 
-	topData, err := getTopData(proCPUFile.Name(), opts.TopFrom, 50)
+	topDataCPU, err := getTopData(proCPUFile.Name(), opts.TopFrom, 50)
+	if err != nil {
+		t.Error(err)
+		return report, err
+	}
+
+	topDataMem, err := getTopData(proMemFile.Name(), opts.TopFrom, 50)
 	if err != nil {
 		t.Error(err)
 		return report, err
@@ -327,7 +339,8 @@ LOOP:
 	report = StressReport{
 		Duration:  duration,
 		Iteration: iteration,
-		Top:       topData,
+		TopCPU:    topDataCPU,
+		TopMem:    topDataMem,
 	}
 
 	if opts.DiffBase != "" {
